refactor(menu card): move into its own namespace folder
